bug fixes / ui improvements / getting rid of deprecated functions
