+++ .astro/settings.json
@@ -1,5 +1,5 @@
 {
 	"_variables": {
-		"lastUpdateCheck": 1742331626662
+		"lastUpdateCheck": 1743375927890
 	}
 }
