+++ package.json
@@ -11,11 +11,11 @@
     "test:cypress": "cypress open"
   },
   "dependencies": {
-    "@astrojs/alpinejs": "^0.4.0",
-    "@astrojs/tailwind": "^5.1.3",
+    "@astrojs/alpinejs": "^0.4.4",
+    "@astrojs/tailwind": "^6.0.2",
     "@atlaskit/pragmatic-drag-and-drop": "^1.5.0",
     "alpinejs": "^3.14.7",
-    "astro": "^5.0.4",
+    "astro": "^5.5.5",
     "js-confetti": "^0.12.0",
     "tailwindcss": "^3.4.16"
   },
