+++ src/components/darkModeToggle.astro
@@ -1,10 +1,4 @@
 ---
-/*
- * Dark Mode Toggle Component
- * Uses Alpine.js store for state management
- * Follows system preference by default
- * Persists user preference in localStorage
- */
 import '../stores/themeStore';
 ---
