+++ src/components/taskManager.astro
@@ -23,6 +23,13 @@ import QuickTaskInput from './QuickTaskInput.astro';
     <div class="flex justify-end mb-6">
         <button 
             x-on:click="$store.todos.deleteCompleted()"
+            x-show="$store.todos.hasCompletedTasks"
+            x-transition:enter="transition ease-out duration-300"
+            x-transition:enter-start="opacity-0 transform scale-90"
+            x-transition:enter-end="opacity-100 transform scale-100"
+            x-transition:leave="transition ease-in duration-200"
+            x-transition:leave-start="opacity-100 transform scale-100"
+            x-transition:leave-end="opacity-0 transform scale-90"
             class="flex items-center gap-2 px-4 py-2 bg-red-50 hover:bg-red-100 dark:bg-red-900/20 dark:hover:bg-red-900/40 text-red-600 dark:text-red-400 rounded-lg transition-colors duration-200"
         >
             <svg class="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
@@ -145,6 +152,8 @@ import QuickTaskInput from './QuickTaskInput.astro';
                                             x-model="task.todo"
                                             x-bind:readonly="task.completed"
                                             x-on:keydown.enter="$event.target.blur()"
+                                            x-on:input="$store.todos.updateTask(task.id, { todo: task.todo })"
+                                            x-on:blur="$store.todos.updateTask(task.id, { todo: task.todo })"
                                             class="flex-1 bg-transparent border-none focus:ring-0 focus:outline-none"
                                             x-bind:class="{
                                                 'line-through text-gray-400 dark:text-gray-500': task.completed,
@@ -244,6 +253,8 @@ import QuickTaskInput from './QuickTaskInput.astro';
                                     x-model="task.todo"
                                     x-bind:readonly="task.completed"
                                     x-on:keydown.enter="$event.target.blur()"
+                                    x-on:input="$store.todos.updateTask(task.id, { todo: task.todo })"
+                                    x-on:blur="$store.todos.updateTask(task.id, { todo: task.todo })"
                                     class="flex-1 bg-transparent border-none focus:ring-0 focus:outline-none"
                                     x-bind:class="{
                                         'line-through text-gray-400 dark:text-gray-500': task.completed,
